move decorator to utils file
